requirements.txt
flask
scikit-learn
opencv-python
numpy
pandas
joblib
pdf2image
python-docx
Pillow
